Fix simplex conversion warnings

+++ src/Inferno/OpenSimplex2.h
@@ -1,8 +1,8 @@
 #pragma once
 
 namespace OpenSimplex2 {
-    float Noise2(long seed, double x, double y);
-    float Noise3_UnrotatedBase(long seed, double xr, double yr, double zr);
+    float Noise2(long long seed, double x, double y);
+    float Noise3_UnrotatedBase(long long seed, double xr, double yr, double zr);
     void Init();
 }
